fix: created a events calendar data file and updated the url path
