+++ client/src/data/event-calendar-data.ts
@@ -1,0 +1,215 @@
+export const mockEvents = [
+    {
+      id: 1,
+      title: "Spring Aquarium Festival",
+      description:
+        "Celebrate the arrival of spring with special themed decorations, limited-time fish species, and exciting challenges. Transform your aquarium into a blooming underwater garden!",
+      startDate: "2025-03-31",
+      endDate: "2025-04-29",
+      location: "Main Aquarium",
+      category: "special",
+      featured: true,
+      isRegistered: true,
+      progress: 35,
+      participants: 1248,
+      rewards: ["Exclusive Cherry Blossom Decorations", "Spring-themed Fish Collection", "2,500 Coins"],
+      requirements: ["Level 5+", "Complete Tutorial"],
+    },
+    {
+      id: 2,
+      title: "Legendary Fish Hunt",
+      description:
+        "A rare opportunity to catch legendary fish species that only appear during this special event. Explore different water zones and complete challenges to increase your chances!",
+      startDate: "2025-04-14",
+      endDate: "2025-04-24",
+      location: "Deep Sea Zone",
+      category: "special",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 876,
+      rewards: ["Legendary Fish Egg", "Golden Fishing Net", "1,000 Coins"],
+      requirements: ["Level 10+", "Own at least 5 different fish species"],
+    },
+    {
+      id: 3,
+      title: "Flash Sale: Premium Decorations",
+      description:
+        "Limited-time discounts on premium aquarium decorations. Upgrade your aquarium with exclusive items at special prices!",
+      startDate: "2025-03-25",
+      endDate: "2025-04-25",
+      location: "In-game Store",
+      category: "offer",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 1543,
+      rewards: ["20% Off Premium Items", "Free Mystery Box"],
+      requirements: [],
+    },
+    {
+      id: 4,
+      title: "Summer Splash Season",
+      description:
+        "Dive into summer with tropical-themed challenges and rewards. Collect summer fish species and unlock beach decorations!",
+      startDate: "2025-05-01",
+      endDate: "2025-07-31",
+      location: "All Zones",
+      category: "season",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 0,
+      rewards: ["Beach Umbrella Decoration", "Tropical Fish Collection", "Summer Treasure Chest"],
+      requirements: ["Level 3+"],
+    },
+    {
+      id: 5,
+      title: "Aquarium Design Championship",
+      description:
+        "Show off your creativity in this aquarium design competition. Submit your best aquarium design and vote for your favorites!",
+      startDate: "2025-05-10",
+      endDate: "2025-05-20",
+      location: "Community Hub",
+      category: "tournament",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 0,
+      rewards: [
+        "Champion Trophy Decoration",
+        "Exclusive Designer Badge",
+        "5,000 Coins",
+        "Featured Spot in Community Showcase",
+      ],
+      requirements: ["Level 8+", "Own at least 10 decorations"],
+    },
+    {
+      id: 6,
+      title: "Breeding Bonanza Weekend",
+      description:
+        "Increased breeding success rates and reduced breeding times for all fish species. Perfect time to expand your collection!",
+      startDate: "2025-05-17",
+      endDate: "2025-05-19",
+      location: "Breeding Pools",
+      category: "special",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 0,
+      rewards: ["Breeding Speed Boost", "Rare Fish Egg", "Premium Fish Food"],
+      requirements: ["Level 7+", "Own at least 2 compatible fish"],
+    },
+    {
+      id: 7,
+      title: "Mystery of the Deep",
+      description:
+        "Explore the mysterious depths of the ocean and uncover hidden treasures. Solve puzzles and complete challenges to reveal the secrets of the deep!",
+      startDate: "2025-04-05",
+      endDate: "2025-04-15",
+      location: "Abyssal Zone",
+      category: "special",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 654,
+      rewards: ["Mysterious Artifact Decoration", "Deep Sea Creature", "1,500 Coins"],
+      requirements: ["Level 12+", "Complete 'Into the Deep' quest"],
+    },
+    {
+      id: 8,
+      title: "Anniversary Celebration",
+      description:
+        "Join us in celebrating our game's anniversary with special events, gifts, and challenges. Thank you for being part of our community!",
+      startDate: "2025-07-15",
+      endDate: "2025-07-25",
+      location: "All Zones",
+      category: "special",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 0,
+      rewards: ["Anniversary Cake Decoration", "Commemorative Fish Species", "Anniversary Gift Box", "3,000 Coins"],
+      requirements: [],
+    },
+    {
+      id: 9,
+      title: "Coral Reef Restoration",
+      description:
+        "Help restore virtual coral reefs and learn about conservation efforts. For each completed challenge, we donate to real-world coral reef conservation!",
+      startDate: "2025-03-15",
+      endDate: "2025-03-30",
+      location: "Coral Reef Zone",
+      category: "special",
+      featured: false,
+      isRegistered: true,
+      progress: 100,
+      participants: 2187,
+      rewards: ["Healthy Coral Decoration Set", "Conservation Badge", "Reef Guardian Title"],
+      requirements: ["Level 3+"],
+    },
+    {
+      id: 10,
+      title: "Premium Membership Discount",
+      description: "Special limited-time offer on Premium Membership. Enjoy exclusive benefits at a reduced price!",
+      startDate: "2025-04-01",
+      endDate: "2025-04-10",
+      location: "In-game Store",
+      category: "offer",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 432,
+      rewards: ["30% Off Premium Membership", "Exclusive Premium Decoration"],
+      requirements: [],
+    },
+    // Adding 3 past events
+    {
+      id: 11,
+      title: "Winter Wonderland",
+      description:
+        "Experience a magical underwater winter with snow-themed decorations, ice fish species, and festive challenges!",
+      startDate: "2024-12-15",
+      endDate: "2025-01-10",
+      location: "Main Aquarium",
+      category: "special",
+      featured: false,
+      isRegistered: true,
+      progress: 100,
+      participants: 3245,
+      rewards: ["Snowflake Decorations", "Arctic Fish Collection", "Ice Castle Centerpiece", "1,500 Coins"],
+      requirements: ["Level 4+"],
+    },
+    {
+      id: 12,
+      title: "Tropical Fish Tournament",
+      description:
+        "Compete with other players to collect and showcase the most vibrant tropical fish species. Win exclusive prizes!",
+      startDate: "2024-09-05",
+      endDate: "2024-09-20",
+      location: "Tropical Zone",
+      category: "tournament",
+      featured: false,
+      isRegistered: true,
+      progress: 100,
+      participants: 1876,
+      rewards: ["Golden Tropical Trophy", "Rare Angelfish", "Tournament Champion Badge"],
+      requirements: ["Level 8+", "Own at least 3 tropical fish species"],
+    },
+    {
+      id: 13,
+      title: "Halloween Spooktacular",
+      description:
+        "Decorate your aquarium with spooky themes and collect ghost fish that only appear during this haunted season!",
+      startDate: "2024-10-20",
+      endDate: "2024-11-02",
+      location: "All Zones",
+      category: "special",
+      featured: false,
+      isRegistered: false,
+      progress: 0,
+      participants: 2543,
+      rewards: ["Haunted Shipwreck Decoration", "Ghost Fish Species", "Pumpkin Treasure Chest"],
+      requirements: ["Level 5+"],
+    },
+  ]
